# Conflicts:
#	ta_app/migrations/0001_initial.py
#	ta_app/models.py
#	ta_app/views/SectionView.py
#	ta_app/views/createCourse.py
#	ta_project/urls.py
#	templates/create_section.html

+++ ta_app/views/deleteCourse.py
@@ -5,6 +5,13 @@ from django.contrib import messages
 
 class deleteCourse(View):
     def get(self, request, course_id):
+        role = request.session['role'];
+        if role != "Admin":
+            return redirect('courseList')
+        if 'role' not in request.session or request.session['role'] != 'Admin':
+            messages.error(request, "You are not authorized to view this page.")
+            return redirect('courseList')
+
         course = Course.objects.filter(pk=course_id).first()
         if course:
             return render(request, 'delete_course.html', {'course': course})
@@ -13,12 +20,15 @@ class deleteCourse(View):
             return redirect('courseList')
 
     def post(self, request, course_id):
-        # Perform the actual deletion if confirmed
-        if 'confirm' in request.POST:
-            course = Course.objects.filter(pk=course_id).first()
-            if course:
-                course.delete()
-                messages.success(request, 'Course deleted successfully!')
-            else:
-                messages.error(request, 'Course not found!')
-        return redirect('courseList')
+        if 'role' in request.session and request.session['role'] == 'Admin':
+            if 'confirm' in request.POST:
+                course = Course.objects.filter(pk=course_id).first()
+                if course:
+                    course.delete()
+                    messages.success(request, 'Course deleted successfully!')
+                else:
+                    messages.error(request, 'Course not found!')
+            return redirect('courseList')
+        else:
+            messages.error(request, "Unauthorized attempt to delete a course.")
+            return redirect('courseList')
